Component tree reclassification
