+++ frontend/app/(main)/page.module.scss
