+++ frontend/app/(signage)/signage/page.tsx
@@ -4,12 +4,12 @@ export const dynamic = "force-dynamic";
 
 import { Suspense, useEffect, useLayoutEffect, useMemo, useRef, useState } from "react";
 import { useSearchParams } from "next/navigation";
-import Logo from "../../components/navigation/navbar/logo";
+import Logo from "../../components/atoms/Logo";
 import CalendarNavbar from "../../components/organisms/CalendarNavbar";
 import DailyView, { defaultRooms } from "../../components/organisms/DailyView";
 import WeeklyView from "../../components/organisms/WeeklyView";
 import { parseSignageFilters, parseSignageView } from "../../../util/signageFilters";
-import navbarStyles from "../../../styles/Navbar.module.scss";
+import navbarStyles from "../../../styles/components/organisms/AppNavbar.module.scss";
 
 const REFRESH_INTERVAL_MS = 2 * 60 * 1000;
 const MIDNIGHT_CHECK_INTERVAL_MS = 30 * 1000;
