+++ frontend/app/components/atoms/BoxText.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from '../../../../styles/components/atoms/BoxText.module.scss';
+import styles from '../../../styles/components/atoms/BoxText.module.scss';
 
 interface BoxProps {
   boxType: 'Meeting Block' | 'Room Block';
