+++ frontend/app/components/atoms/CheckBox.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from "../../../../styles/components/atoms/Checkbox.module.scss"
+import styles from "../../../styles/components/atoms/Checkbox.module.scss"
 
 interface LabeledCheckBoxProps {
     label: string;
