--- frontend/app/components/navigation/navbar/button.jsx
@@ -1,7 +1,0 @@
-const Button = () => {
-  return (
-    <button className="h-12 rounded-lg bg-white font-bold px-5">Sign In</button>
-  );
-};
-
-export default Button;
