+++ frontend/app/components/organisms/NewMeeting.tsx
@@ -5,8 +5,8 @@ import TextField from '../atoms/TextField';
 import ModeTypeButtons from '../atoms/ModeTypeButtons';
 import DatePicker from '../atoms/DatePicker';
 import TimePicker from '../atoms/TimePicker';
-import Dropdown from '../atoms/dropdown';
-import LabeledCheckbox from '../atoms/checkbox';
+import Dropdown from '../atoms/Dropdown';
+import LabeledCheckbox from '../atoms/CheckBox';
 import RecurringMeetingForm from '../molecules/RecurringMeeting';
 import CloseIcon from '@mui/icons-material/Close';
 import IconButton from '@mui/material/IconButton';
