--- frontend/styles/components/atoms/SolidButton.module.scss
@@ -1,14 +1,0 @@
-.solidbutton {
-  border: none;
-  color: white;
-  padding: 10px 20px;
-  font-size: 16px;
-  border-radius: 8px;
-  cursor: pointer;
-  margin-bottom: 5px;
-  margin-top: 5px;
-}
-
-.solidbutton:hover {
-  opacity: 0.9;
-}
