--- frontend/styles/components/atoms/UploadPandaDocs.module.scss
@@ -1,123 +1,0 @@
-.uploadLabel {
-  font-family: 'Source Sans Pro', sans-serif;
-  font-size: 18px;
-  margin-bottom: 15px;
-}
-
-.uploaddroparea {
-  width: 95%;
-  height: 150px;
-  border: 2px dashed gray;
-  border-radius: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-}
-
-.uploadplaceholder {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: #555;
-  margin-bottom: -35px;
-  cursor: pointer;
-
-  svg {
-    margin-top: 10px;
-    scale: 4;
-    color: gray;
-    margin-bottom: 25px;
-  }
-
-  .uploadbutton {
-    color: rgb(57, 48, 48);
-    border-radius: 5px;
-  }
-
-  .fileinput {
-    display: none;
-  }
-
-  .underline {
-    text-decoration: underline;
-  }
-}
-
-.uploadfileinfo {
-  display: flex;
-  flex-direction: column;
-
-  p {
-    font-size: 16px;
-    color: #333;
-  }
-
-  .progressbarcontainer {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    width: 100%;
-  }
-
-  .progressbar {
-    width: 100%;
-    height: 8px;
-    background-color: #484848;
-    border-radius: 4px;
-    transition: width .4s ease;
-  }
-
-  .progresspercentage {
-    font-size: 14px;
-    color: #666666;
-    margin-left: auto;
-  }
-
-  .uploadfileinfo {
-    display: flex;
-    flex-direction: row;
-    gap: 50px;
-
-    svg {
-      color: gray;
-      margin-top: auto;
-      margin-bottom: auto;
-      scale: 2;
-      padding: 8px;
-      box-shadow: 0 0 1px lightgray;
-      border-radius: 7px;
-    }
-
-    .filename {
-      font-weight: 600;
-      font-size: 14px;
-      text-align: left;
-      color: #000000;
-      margin-bottom: 0px;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      max-width: 200px;
-    }
-
-    .filesize {
-      font-weight: 400;
-      font-size: 14px;
-      text-align: left;
-      color: #666666;
-      margin-top: 5px;
-    }
-  }
-
-  .cancelupload {
-    font-weight: 900;
-    font-size: 20px;
-    color: #878787;
-    cursor: pointer;
-  }
-}
-
-.uploaded {
-  border: solid 2px #CC3366;
-}
