+++ frontend/styles/components/organisms/AppNavbar.module.scss
