+++ frontend/styles/components/organisms/CalendarNavbar.module.scss
@@ -73,7 +73,7 @@
     font-family: 'Source Sans Pro', sans-serif;
     appearance: none;
     padding-right: 24px;
-    background-image: url('/drop-down-arrow.svg');
+    background-image: url('/svg/drop-down-arrow.svg');
     background-repeat: no-repeat;
     background-position: right 10px center;
   }
